+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+- Added function `init()` to the API to initialize a Git. This function prompt for confirmation if git is already initialized.
+
 ## [1.1.1] - 2023-03-22
 
 ### Fixed